+++ Tests/TPPDFIntegrationTests/ExternalDocumentTests.swift
@@ -4,14 +4,10 @@ import Nimble
 @testable import TPPDF
 import PDFKit
 
+@available(iOS 11.0, *)
 class ExternalDocumentTests: QuickSpec {
 
     override func spec() {
-        // This is a hotfix until we can use xcodebuild with SPM resource paths
-        if Bundle.module == Bundle.main {
-            Bundle.module = Bundle(for: ExternalDocumentTests.self)
-        }
-        
         describe("External Document") {
             context("only document") {
                 context("debug mode") {